KAIZEN-0: Flytt dialogcontainer i egen fil
