+++ src/App.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
-import { BrowserRouter as Router, matchPath, Route, RouteComponentProps, Switch, withRouter } from 'react-router-dom';
+import { BrowserRouter as Router, matchPath, RouteComponentProps, withRouter } from 'react-router-dom';
 import { Provider } from './Context';
 import { DialogOverview } from './view/DialogOverview';
-import Dialog from './view/Dialog';
-import DialogNew from './view/DialogNew';
 import { AlertStripeContainer } from './view/AlertStripeContainer';
 
 import './App.less';
 import classNames from 'classnames';
 import { DialogBanner } from './view/DialogBanner';
 import AktivitetContainer from './view/AktivitetContainer';
+import DialogContainer from './view/DialogContainer';
 
 function Routing() {
     return (
@@ -26,18 +25,6 @@ function Routing() {
 export default Routing;
 
 interface Props extends RouteComponentProps<{ dialogId?: string }> {}
-
-function DialogContainer(props: { className: string }) {
-    return (
-        <div className={props.className}>
-            <Switch>
-                <Route exact path="/" component={Dialog} />
-                <Route path="/ny" component={DialogNew} />
-                <Route path="/:dialogId" component={Dialog} />
-            </Switch>
-        </div>
-    );
-}
 
 function App(props: Props) {
     const harDialogId = matchPath<{ dialogId: string }>(props.location.pathname, '/:dialogId');
